requirements.txt
pyTelegramBotAPI==4.22.1
loguru==0.7.2
peewee==3.17.6
pydantic-settings==2.4.0
python-dotenv==1.0.1
python-telegram-bot-pagination==0.0.3
requests==2.32.3